Fixes tests

+++ database/migrations/2019_09_01_175344_drop_slug_column_from_tags.php
@@ -13,9 +13,11 @@ class DropSlugColumnFromTags extends Migration
      */
     public function up()
     {
-        Schema::table('tags', function (Blueprint $table) {
-            $table->dropColumn('slug');
-        });
+        if (Schema::hasColumn('tags', 'slug')) {
+            Schema::table('tags', function (Blueprint $table) {
+                $table->dropColumn('slug');
+            });
+        }
     }
 
     /**
